requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
djangorestframework==3.12.4
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
joblib==1.0.1
MarkupSafe==1.1.1
nltk==3.6.1
numpy==1.20.2
packaging==20.9
pandas==1.2.4
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
scikit-learn==0.24.1
scikit-multilearn==0.2.0
scipy==1.6.2
six==1.15.0
sqlparse==0.4.1
threadpoolctl==2.1.0
tqdm==4.60.0
uritemplate==3.0.1
urllib3==1.26.4
